backend/requirements.txt
amqp==5.0.1
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
billiard==3.6.3.0
black==20.8b1
boto3==1.16.4
botocore==1.19.4
celery==5.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coverage==5.3
cryptography==3.2
dateparser==1.0.0
decorator==4.4.2
Deprecated==1.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-coverage-plugin==1.8.0
django-environ==0.4.5
djangorestframework==3.12.1
docker==4.3.1
flake8==3.8.4
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
isort==5.6.4
jmespath==0.10.0
kombu==5.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyGithub==1.53
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.1.1
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-flake8==1.0.6
pytest-json-report==1.2.4
pytest-metadata==1.10.0
python-dateutil==2.8.1
python-telegram-bot==13.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.23
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
wrapt==1.12.1
zipp==3.3.2
zope.event==4.5.0
zope.interface==5.1.2
